junk/Memory/requirements.txt
google-genai
pyaudio